remove from tasks and rename flags
