+++ authentik/enterprise/audit/apps.py
@@ -3,8 +3,8 @@
 from django.conf import settings
 from django.utils.translation import gettext_lazy as _
 
+from authentik.admin.flags import Flag
 from authentik.enterprise.apps import EnterpriseConfig
-from authentik.tenants.flags import Flag
 
 
 class AuditIncludeExpandedDiff(Flag[bool], key="enterprise_audit_include_expanded_diff"):
