+++ authentik/flows/apps.py
@@ -3,9 +3,9 @@
 from django.utils.translation import gettext_lazy as _
 from prometheus_client import Gauge, Histogram
 
+from authentik.admin.flags import Flag
 from authentik.blueprints.apps import ManagedAppConfig
 from authentik.lib.utils.reflection import all_subclasses
-from authentik.tenants.flags import Flag
 
 GAUGE_FLOWS_CACHED = Gauge(
     "authentik_flows_cached",
@@ -25,7 +25,6 @@ HIST_FLOWS_PLAN_TIME = Histogram(
 
 
 class RefreshOtherFlowsAfterAuthentication(Flag[bool], key="flows_refresh_others"):
-
     default = False
     visibility = "public"
     description = _("Refresh other tabs after successful authentication.")
@@ -33,7 +32,6 @@ class RefreshOtherFlowsAfterAuthentication(Flag[bool], key="flows_refresh_others
 
 
 class ContinuousLogin(Flag[bool], key="flows_continuous_login"):
-
     default = False
     visibility = "public"
     description = _("Upon successful authentication, re-start authentication in other open tabs.")
